+++ src/hatch_fancy_pypi_readme/_builder.py
@@ -5,11 +5,19 @@
 from __future__ import annotations
 
 from ._fragments import Fragment
+from ._substitutions import Substituter
 
 
-def build_text(fragments: list[Fragment]) -> str:
+def build_text(
+    fragments: list[Fragment], substitutions: list[Substituter]
+) -> str:
     rv = []
     for f in fragments:
         rv.append(f.render())
 
-    return "".join(rv)
+    text = "".join(rv)
+
+    for sub in substitutions:
+        text = sub.substitute(text)
+
+    return text